# Reduce the size of Box.module.css