--- a/packages/react/src/Box/Box.module.css
+++ b/packages/react/src/Box/Box.module.css
@@ -1,4 +1,13 @@
 .Box {
+  --box-npbs: 0;
+  --box-npbe: 0;
+  --box-npis: 0;
+  --box-npie: 0;
+  --box-nmbs: 0;
+  --box-nmbe: 0;
+  --box-nmis: 0;
+  --box-nmie: 0;
+
   padding-block: var(--box-npbs) var(--box-npbe);
   padding-inline: var(--box-npis) var(--box-npie);
 
@@ -8,6 +17,15 @@
 
 @media screen and (min-width: 48rem) {
   .Box {
+    --box-rpbs: 0;
+    --box-rpbe: 0;
+    --box-rpis: 0;
+    --box-rpie: 0;
+    --box-rmbs: 0;
+    --box-rmbe: 0;
+    --box-rmis: 0;
+    --box-rmie: 0;
+
     padding-block: var(--box-rpbs) var(--box-rpbe);
     padding-inline: var(--box-rpis) var(--box-rpie);
 
@@ -18,6 +36,15 @@
 
 @media screen and (min-width: 80rem) {
   .Box {
+    --box-wpbs: 0;
+    --box-wpbe: 0;
+    --box-wpis: 0;
+    --box-wpie: 0;
+    --box-wmbs: 0;
+    --box-wmbe: 0;
+    --box-wmis: 0;
+    --box-wmie: 0;
+
     padding-block: var(--box-wpbs) var(--box-wpbe);
     padding-inline: var(--box-wpis) var(--box-wpie);