… .changeset/hot-moose-attack.md
… packages/react/src/Box/Box.module.css.d.ts