The CSS for the Box component was previously ~2,800 lines of code. This PR reduces it to 135 lines.

The result of this is a 22% reduction in the size of our main CSS bundle, and an 18% reduction in the size of our main JavaScript bundle.

|   | Size Before | Size After | Size Change | % Change |
|-------------|------------:|-----------:|------------:|---------:|
| `main.css`  | 629 KB      | 497 KB     | -132 KB     | -22%     |
| `index.js`  | 429 KB      | 351 KB     | -78 KB      | -18%     |
| **`total`** | **1058 KB** | **848 KB** | **-210 KB** | **-20%** |

There is a trade-off in that the rendered HTML markup for `Box` may now be slightly longer due to the addition of inline CSS variables when compared to the best-case scenario from the previous implementation. When compared to the worst-case scenario, however, it's about half of the size.

```html
<!-- Before (best case) -->
<div class="Box-module__Box-padding--80__fB_np">
</div>

<!-- Before (worst case) -->
<div class="Box-module__Box-narrow-paddingBlockStart--8__lH0Nx Box-module__Box-regular-paddingBlockStart--12__gOq9H Box-module__Box-wide-paddingBlockStart--16__f5Dfj Box-module__Box-narrow-paddingInlineEnd--8__Rt9bb Box-module__Box-regular-paddingInlineEnd--12__X02EG Box-module__Box-wide-paddingInlineEnd--16__K4CHM Box-module__Box-narrow-paddingBlockEnd--8__y1kaa Box-module__Box-regular-paddingBlockEnd--12__uroYu Box-module__Box-wide-paddingBlockEnd--16__yvgLa Box-module__Box-narrow-paddingInlineStart--8__WQHFH Box-module__Box-regular-paddingInlineStart--12__MFP_O Box-module__Box-wide-paddingInlineStart--16__vDqYc Box-module__Box-narrow-marginBlockStart--8__i685k Box-module__Box-regular-marginBlockStart--12__HLits Box-module__Box-wide-marginBlockStart--16__R8o04 Box-module__Box-narrow-marginInlineEnd--8__jhqqu Box-module__Box-regular-marginInlineEnd--12__VJwpo Box-module__Box-wide-marginInlineEnd--16__JaWXe Box-module__Box-narrow-marginBlockEnd--8__kUGhn Box-module__Box-regular-marginBlockEnd--12__kpsJs Box-module__Box-wide-marginBlockEnd--16__LqPgv Box-module__Box-narrow-marginInlineStart--8__Md4jm Box-module__Box-regular-marginInlineStart--12__LVaCT Box-module__Box-wide-marginInlineStart--16__mUIjr">
</div>

<!-- After (fixed) -->
<div class="Box-module__Box__iXAWK" style="--box-npbs: var(--base-size-80); --box-rpbs: var(--base-size-80); --box-wpbs: var(--base-size-80); --box-npbe: var(--base-size-80); --box-rpbe: var(--base-size-80); --box-wpbe: var(--base-size-80); --box-npis: var(--base-size-80); --box-rpis: var(--base-size-80); --box-wpis: var(--base-size-80); --box-npie: var(--base-size-80); --box-rpie: var(--base-size-80); --box-wpie: var(--base-size-80); --box-nmbs: 0; --box-rmbs: 0; --box-wmbs: 0; --box-nmbe: 0; --box-rmbe: 0; --box-wmbe: 0; --box-nmis: 0; --box-rmis: 0; --box-wmis: 0; --box-nmie: 0; --box-rmie: 0; --box-wmie: 0;">
</div>
```

Short variable names have been chosen to reduce the impact of this. For example, `--box-npie` refers to a `Box` component on a ***N***arrow viewport and sets the ***P***adding-***I***nline-***E***nd property.


## Backwards compatibility

I'm pretty confident that the underlying logic hasn't changed and the VRT snapshots suggest the same. However please do let me know if you think there's something I've missed.

## Tests

The spacing variables have quite a few tests associated with them — _1008_ to be exact. The test suite takes advantage of Jest's `.each()` function to test all of the variables, for all of the spacing props, against all of the available options (`none`, `condensed`, `4`, `8`, etc). It may seem like overkill, but the benefit of this approach is that it gives really clear messages when _something_ breaks.

![image](https://github.com/user-attachments/assets/b49c5c95-6b22-4a8b-9433-ed1090998a85)

For example, if a test fails you will see:

```
when marginBlockStart > equals 8 > --box-rmbs is not being set to var(--base-size-8)
```

[You can see the test in question here](https://github.com/primer/brand/blob/612f1509e28b129d5e7bd228a26fef456f0551cd/packages/react/src/Box/Box.test.tsx#L127-L138)

The benefit of this approach over the established pattern of running a `for/of` loop inside of the test is that each assertion is part of its own test, which isn't the case when using a `for/of` loop. When an assertion fails you'll see exactly which test fails, as shown above. If we were to use a `for/of` loop, when an assertion fails the whole test will fail, and we won't get the same level of detail in the test output.

### Test performance

Running tests is pretty fast. Running all 1008 tests takes <3.5 seconds on my machine. Looking at [recent runs of the CI / Install and run tests Action](https://github.com/primer/brand/actions/workflows/ci.yml?query=is:success) it looks like the impact of these additional tests is pretty negligible.

The number of tests could be reduced by reducing the number of spacing props we're testing against. Instead of testing against all of our valid base scale values (4, 8, 12, 16, etc) we could just pick one number to test along with our string spacings (`'none', 'condensed', 'normal', 'spacious'`). This would bring the number of tests down to 240, however it would only reduce the time taken to run the tests by ~10%.


## What should reviewers focus on?

- Happy with the implementation?
- Happy with the tests?
- Can you find any visual regressions, or reasons that the underlying logic might have changed?


## Contributor checklist:

- [x] All new and existing CI checks pass
- [x] Tests prove that the feature works and covers both happy and unhappy paths
- [x] Any drop in coverage, breaking changes or regressions have been documented above
- [x] New visual snapshots have been generated / updated for any UI changes
- [x] All developer debugging and non-functional logging has been removed
- [x] Related issues have been referenced in the PR description

## Reviewer checklist:

- [ ] Check that pull request and proposed changes adhere to our [contribution guidelines](../../CONTRIBUTING.md) and [code of conduct](../../CODE_OF_CONDUCT.md)
- [ ] Check that tests prove the feature works and covers both happy and unhappy paths
- [ ] Check that there aren't other open Pull Requests for the same update/change