@joshfarrant 2024-09-04
Thanks for the review @rezrah, appreciate you taking the time to dig into it.

One thing I forgot to note in the PR description is that there's already precedent for a similar pattern within the Section component.

https://github.com/primer/brand/blob/main/packages/react/src/Section/Section.tsx#L123-L133

https://github.com/primer/brand/blob/main/packages/react/src/Section/Section.module.css#L8-L15

<img width="1281" alt="image" src="https://github.com/user-attachments/assets/24f3e112-e436-4b4c-bdba-b5ef71d5ad9e">

```html
<section
  style="--brand-Section-background-image-position: 50%; --brand-Section-background-image-size: cover;"
/>
```

It's a bit less pronounced here as we (hopefully 😅) don't have to worry about consumers nesting Sections within Sections, and so we don't have to reset unused values as I do with Box. We're also likely to only have a few Section components on a page — compared to a lot of Box components — so there's less pressure to reduce the size of the CSS variable names.