@joshfarrant 2024-09-04
@rezrah Just to note, I dug into this a bit more and managed to find a better solution to resetting the CSS variables, and it turns out we don't need `@property` at all as I previously thought 🎉

Take a look at 6f6574de8b7b7d7ba38daf093bfa965a6253af6a, it's actually more straightforward than I initially thought.

This change removes the need for all of the `--box-npie: 0; --box-rpie: 0;` resets and reduces the size of the markup